Move types out of separate file
